requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
pysrt==1.1.2
